lots of fixes/new instructions

+++ src/util/types.h
@@ -5,7 +5,7 @@
 
 typedef uint8_t u8;
 typedef uint16_t u16;
-typedef uint16_t u32;
+typedef uint32_t u32;
 
 typedef int8_t s8;
 typedef int16_t s16;
